Add files via upload
The Title of our project model is “ATM Simulator System”, the project is developed in JAVA language . “ATM Simulator System” project is model that enables the customer to perform the basic banking transactions by sitting at home or at their office through PC or Laptop. The system provides the customer the facility to first create an account in the bank if he is not the registered the customer of the bank. 
                  For account creation of the customer his personal information is taken through three separate pages each regarding to separate types of user information, these details are then stored in the database. At the end of the account creation process, a unique account number and pin is created for every customer, pin can be changed later on.
              If the user is already registered he can use his unique Account number and PIN to sign up into the ATM System to perform various actions, transactions and view Mini statements through the ATM.
The user can perform the transactions on account as per their requirement , but following are the transactions and actions which can be performed :
Deposit, Withdrawl, Fastcash, BalanceEnquiry, PinChange and generating Mini Statement .